[2/2] switch to infer_dense_chords_for_sequence
Switch from chord_inference.infer_chords_for_sequence, which requires beat annotations, to sequences_lib.infer_dense_chords_for_sequence, which adds one chord for each time one of a group of simultaneously played notes changes.

Note that this is highly imperfect as 1) it won't give any chords for single line melodys that lack harmony, 2) it'll give overly complicated chords when it captures a state with passing tones, and 3) it doesn't attempt to do any sort of "smoothing" of chords across a beat or larger chunk of measure.

Later code will attempt to rectify (2) and (3), but (1) is beyond the scope of this relatively naive approach to chord annotation.