Add chords onto Notes instead of TextAnnotation
Add a param in sequences_lib.infer_dense_chords_for_sequence to add_chords_to_notes, ie directly add chord root and quality to each Note submessage instead of as unattached TextAnnotations. This will help significantly in simplifying how to load the data for the Piano Genie model.

Note: this approach may result in (-1, -1) being set for the first few chords if a chord isn't immediately determined. Future implementation should change this to crop the overall NoteSequence when this is the case.